backend/requirements.txt
fastapi
uvicorn
sqlite-utils
sqlalchemy
aiosqlite
slowapi
pydantic
python-multipart
jinja2
email-validator
python-dotenv